Start populating the handbook with useful cross reference targets.
All cross reference labels start with name of the file that contains
them.  A label for the top section level is simply the name of the
file (omitting the .sgml).  Other references within the file append a
colon and onother name.  For example, the label on the mailing list
section in the file eresources.sgml is eresources:mail.  This gives
each file its own cross reference namespace.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@22 c2e8774f-c49f-e111-b436-862b2bbc8956