Allow DOCS to be overriden and conditionally clear the SUBDIR.
This is to overcome a problem in building this part of the tree
from the new build structure. The current system relies on picking
up a Makefile.inc from a parent directory that sets WEBBASE and as
the subdirectories are entered this Makefile.inc is from a different
parent directory with a different WEBBASE setting.

This wasn't something I could workaround when building from the
new directory structure so as a short term cludge I now enter this
directory twice, once with NEW_BUILD defined so the subdirs are
skipped and once with DOCS set to null so only the directories are
built and WEBBASE is set appropriately in each case.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@1553 c2e8774f-c49f-e111-b436-862b2bbc8956