As discussed in -doc, this jumpstarts the doc/en/ hierarchy for documentation in English, to be consistent with the doc/ja/ hierarchy.
Rather than moving the existing FAQ and Handbook over to this directory,
they will be converted to the DocBook DTD, and the results of the conversion
will be committed to this new hierarchy. The existing doc/faq and
doc/handbook can then be removed.

The README file will be used to explain the current state of the DocBook
conversion process for the Handbook, so that people without easy access
to the CVS logs can keep track of what's going on, and lend a hand where
appropriate.

I plan to start converting the Handbook around 1830 BST.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@2627 c2e8774f-c49f-e111-b436-862b2bbc8956