This is a DSSSL stylesheet. It builds upon Norm Walsh's modular DocBook stylesheet (ports/textproc/dsssl-docbook-modular) and extends it in three ways.
First, Norm's stylesheets have a number of documented 'knobs' that can be
twiddled to produce specific effects. For example, one of the options I've
turned on names the generated HTML files after the id attribute of a
section. Say goodbye to c01234.html, and wave hello to 'kerneloptions.html'.

Second, some of Norm's design choices don't follow our existing practice.
For example, we indicate application names in bold, command names in a
monospace font. Norm's sheets just render command names in bold. So we
override his definition with our own for some elements.

Thirdly, we tell the stylesheet about elements that don't exist in DocBook
but do exist in our extended DTD.

Tim Vanderhoek is probably sliding off his seat about now :-)

There's more that needs doing to this, but it's enough to get start with.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@3014 c2e8774f-c49f-e111-b436-862b2bbc8956