<foo        id="bar">        ...
changed to

    <foo id="bar">
       ...

Before people complain that "Hang on, now you can't find out what the
allocated ID values are with a simple 'grep'" I'll say that's not a
problem. I plan to introduce a target in the Makefile (probably
something like 'handbook.id' which will automatically generate this
list doing a proper SGML parse.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@3683 c2e8774f-c49f-e111-b436-862b2bbc8956