Huge whitespace changes. Translators can ignore this commit completely.
Rationale: All the changes to the DocBook handbook so far have been
careful to keep whitespace changes to a minimum. This is so the
translators have as easy a job as possible in identifying exactly what's
changed.

This has meant the English version has become more and more 'ugly'. Lines
indented by the wrong amount, some lines longer than 130 characters,
others shorter than 20, gaps of 3 or 4 lines between paragraphs (and
sometimes within paragraphs). This makes it difficult to follow the
structure of the document, and needlessly complicates fixing SGML
problems.

It also makes the source practically useless as a teaching aid; the
more baroque the source looks, the less likely people are to dive in and
contribute.

This commit fixes all that -- and boy was it tedious. The snag is, it's
touched almost every line in every file in the Handbook.

Technically, the changes were made by running (in Emacs)
sgml-indent-or-tab (bound to the TAB key) on almost each line (except
those in <programlisting>, <screen>, <literallayout>, and other
verbatim sections), and then running sgml-fill-element (bound to
C-c C-q) on most paragraphs.

FWIW, this is the first, only, and last change of this type contemplated.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@4472 c2e8774f-c49f-e111-b436-862b2bbc8956