Add SGML comments at the top of the files with the $Id$ string. Remove most of the other spurious comments.
Two comments relating to copyright have *not* been merged in from the
LinuxDoc version yet -- I've contacted the original authors to ask if
they would be willing to assign the copyright to the project. When I
get their response the copyright comments will either be merged in, or
left out, as necessary.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@4477 c2e8774f-c49f-e111-b436-862b2bbc8956