The Handbook conversion has changed the names of some of the generated HTML files (not all of them, just some of them).
Fix up the broken references to point to the correct files.

This has not touched the release notes in releases/*/*.  I'm not sure
whether they should remain as historical record of what was shipped or
not.  About 5 of them now have a couple of broken links each, which will
be trivial to fix.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@4571 c2e8774f-c49f-e111-b436-862b2bbc8956