Add notes stating which release of FreeBSD each y2k bug was (will be) fixed in.
Wow.  Do you realize this file will become irrelevant in 4 months?  Summer
vacation is almost over.  Two months ago I wrote my last exam.  At the time,
four months seemed like a merciful eternity for a summer vacation.  Now there
are only two months left.  Four months is pretty short.  I remember when the
pr misc/1380 was filed, I read the comment about how the bug reported in
that pr would surface in four years.  My first thought at the time was
"Wow, that's pretty soon."  Come to think of it, in retrospect, it wasn't
even that long ago that I was young enough to suffer the illusion (or "be so
innocent as to believe", YMMV) that the year two-thousand would never arrive.

I'm allowed to waste verbiage in this commit log.  The file is doomed in
four months, anyways.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@5240 c2e8774f-c49f-e111-b436-862b2bbc8956