Miscellaneous small changes.  Includes reformatting white space in some of the example HTML and DocBook fragments, which are significant.
Explain where to find the FreeBSD DTD that extends DocBook.

Add a section explaining about books and articles, and the differences
between them.

Added an example of <procedure>.

Be consistent with the white space inside <programlisting>.  Make sure
that content starts immediately after the opening tag.  Shouldn't make
a great deal of difference, but we should practice what we preach.

Remove the information about <informalexample>, as it seems to be
redundant, and we don't use it anymore.

Correct an error ("]]", should be "]") in the example of including the manual
page reference entities.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@5508 c2e8774f-c49f-e111-b436-862b2bbc8956