Chapter 2 of "The Design and Implementation of the 4.4BSD Operating System"
By Kirk McKusick, Keith Bostic, Michael Karels, and John Quarterman.

Graciously allowed in by those lovely folks at Addison-Wesley.

Heaps of kudos to Sheldon (sheldonh) for doing most of the work to bring
this in, including almost all the markup.  It's been waiting for the
image support to finalise, and now that we've done that we're ready to go.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@8812 c2e8774f-c49f-e111-b436-862b2bbc8956