The problem is with nested links; quite frankly, they don't work. This isn't normally an issue because no one in their right mind will stick a <link> inside another <link>.  However, we have entities which create links, such as &man.*;, &a.*;, etc.  It's nice to be able to use these inside links.  To deal with this..
Introduce a create-link procedure which will be used to replace (make
element gi: "A" ...) constructs.  This procedure creates a link as
specified only if the can-link-here procecure (described below)
returns #t.  If the latter returns #f, it will print the link text
without the link.

The (also new) can-link-here procedure returns #t if it determines
that it's okay to make a link in the current context, and #f
otherwise.  Currently, it does its check by figuring out whether the
current context is within a <title> or <question> tag.  This is not
ideal because it doesn't catch all cases, but it's a lot better than
nothing.  As the other cases are discovered, this procedure can be
modified.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@9528 c2e8774f-c49f-e111-b436-862b2bbc8956