Wrap author attribution in <chapterinfo>.
Reword a couple of run-on sentences to use better subject-verb
agreement.

Terminate all items in a list with a period, not just some of
them.

Add <username> and <application> tags as appropriate.

"pw(8) has the information"

doesn't look well without the benefits of a hyperlink.  Reword this to:

"For more information, please see pw(8)."

Update example to change shells to zsh since tcsh is now in the base
system.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@9953 c2e8774f-c49f-e111-b436-862b2bbc8956