Remove Greg Lehey's X11 section.  It has served us well.
There are some gaps in the content now but those will be filled in
shortly.  As far as I can tell, this was the last part of the Handbook
where the copyright was held by an individual author.  I wonder if
people like iUniverse.com contacted Greg before publishing the
Handbook.

While I'm here add a standard synposis and mark up some attributions
in <sectNinfo>.

Partially submitted by: nik (posting on -doc)


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@10209 c2e8774f-c49f-e111-b436-862b2bbc8956