An increasing amount of documents use the -V flag to Jade via JADEFLAGS to set variables such as %generate-article-toc%.  However, JADEFLAGS is also passed to nsgmls, which doesn't, and shouldn't, understand -V.  The Makefiles which do this are correct, because the name--JADEFLAGS--implies that it will only be passed to Jade, not to nsgmls, too.  Furthermore, simply not passing JADEFLAGS to nsgmls is not okay, since nsgmls *does* need the -i flags used to include/exclude certain parts of the document.
Remedy this by breaking up JADEFLAGS into itself and SGMLFLAGS.  The
latter will be passed to all SGML processors such as nsgmls and Jade.
The former will only be passed to Jade.  The -V flags should stay in
JADEFLAGS, and the -i flags should be moved to SGMLFLAGS.

This fixes `make lint` for documents which use -V via JADEFLAGS.

Reviewed by:	bmah


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@10372 c2e8774f-c49f-e111-b436-862b2bbc8956