Rework our index support.
  * Add a new document-specific variable, HAS_INDEX, to specify if a
    given document is marked up with <indexterm> entries.

  * Rework the index support so that both HAS_INDEX and GEN_INDEX are
    checked before trying to generate an index for a document.

  * Only create index.sgml if both HAS_INDEX and GEN_INDEX are set.

This allows us to recursively build the documentation tree with
GEN_INDEX=1 and have it only try to create an index (very time
consuming) for the few documents that are ready for this.  Previously,
running "make GEN_INDEX=1" from the top of the doc tree would look for
index terms in every single document.

With this, I hope we can start building our docs with GEN_INDEX set on
freefall so that users browsing the HTML docs will get the benefit of
the index we've been hiding in CVS for 6 months.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@11247 c2e8774f-c49f-e111-b436-862b2bbc8956