Be pedantic:
"A port is not the minimal set of files to compile a program.  The
minimal set of files needed to compile a program includes the source
code.  It's a minimal set of files that tells your system how to
compile that source code."

PR:		docs/32109
Submitted by:	Michael Lucas <mwlucas@blackhelicopters.org>


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@11259 c2e8774f-c49f-e111-b436-862b2bbc8956