Based on some verbal feedback from colleagues;
De-emphasise the fact that we use GNATS.  Some people have assumed that
GNATS is something they have a problem with, not a tool that we use.

Add some additional information about what a PR is, and why it's not
necessarily a bad thing.

Include an additional link to send-pr.html where it seemed appropriate.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@13367 c2e8774f-c49f-e111-b436-862b2bbc8956