Introduce support for Mandatory Access Control and extensible kernel access control.
Document the kernel side of the MAC policy architecture.  This
is a little out of date at the moment.  Some parts to be filled
in as things are developed, and much is subject to change.  It
will, however, give developers a good idea of how things work.

Obtained from:	TrustedBSD Project
Sponsored by:	DARPA, NAI Labs


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@13938 c2e8774f-c49f-e111-b436-862b2bbc8956