Fix a long-standing bug in the way (the lack of) images is handled in the 'text-only' case.
Do this by explicitly setting preferred-mediaobject-notations to the
empty list.

This has only worked for so long because the source format for images in
all the documents with plain text equivalents for images has been EPS,
which isn't used in the HTML output format.  If PNG images were used
then the <textobject> element wasn't chosen.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@14779 c2e8774f-c49f-e111-b436-862b2bbc8956