Add a project status page ala C99 and similar for FreeBSD/MIPS, to make it easier for people to track the status.
Add a developers.sgml entry for Justin Wojdacki for use in the status page.

Reviewed by:	ceri

Add an entry for MIPS on the platforms page, and a small page there with
a few tiny bits of information, mostly mailing list info and a link to
the afforementioned status page.

Requested by:	ceri


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@14871 c2e8774f-c49f-e111-b436-862b2bbc8956