Update the 5.0-RELEASE task list:
Remove X11 instability issue, it was fixed by peter, bde, and others as
part of ABI and signal compat.

Replace DP2 requirements section with an RC1 requirements section.

Add a final release requirement that we tone down the debugging
options.

Remove the TrustedBSD MAC framework system call interface item: it's
done.

Remove the GEOM as default storage management framework: it's done.

Update the KerberosV/OpenPAM issue: nectar fixed bugs in the PAM module
with the help (hinderance) of phkmalloc.  Some bugs remain in the
OpenSSH integration area.

Add a new item for the creation of a freebsd-games port, requested
by kkenn.  Assign markm as owner.

Add a new item for the creation of an a.out buildchain port, requested
by kkenn.

DEVELOPERS: please review the todo list and make sure it accurately
captures the list of "must be done" things for 5.0-RELEASE.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@14943 c2e8774f-c49f-e111-b436-862b2bbc8956