Rename 'Status' to 'Tasks'.
Add a new "Subsystems and Staffing" section to identify some of the major
subsystems, their SMPng status, and current developers working on them.
This is not intended to be a complete list, I may well have left off
several substantial sub-systems, important contributions, etc.  If so,
please let me know.  The goal is to identify at a higher level what is
being done, what is planned to be done, and what has been done so that
developers interested in tackling a chunk of the work larger than a
small task can do so more easily.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@15924 c2e8774f-c49f-e111-b436-862b2bbc8956