Add an event section to the website.
This is meant both as a way to get an overview of upcoming FreeBSD
related events, and as a way to structure links to pages with
information / pictures / etc. from past events.

The page automatically determines which events are upcoming and which
are past, by comparing the event dates with the current date.

Discussed on:		-www
Inspired by PR:		www/53674
Which was submitted by:	Josef El-Rayes <j.el-rayes@daemon.li>


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@18407 c2e8774f-c49f-e111-b436-862b2bbc8956