[ports committer] Rearrange the FreeBSD Internal web pages to accomplish several goals:
1. Make the pages more consistent with the style of the other FreeBSD
www pages.
2. Reorganize the top-level Internal page into four subheadings:
Projects, Policies, People, and Resources.  This helps make the
page 'flow' a lot better.
3. Create a new, second-level, Policies page to hold one subgrouping
of the links that used to live on the main page, which had too many
links as it was.
4. Remove the Contacting, Copyright, and Search links, which had
been promoted to the top-level FreeBSD page some time ago in another
cleanup.

Note: all of the links (except in #4) from the original page remain
either in pages committed -- this is essentially a reorganization of
the existing links to present them in a more logical -- and legible --
context.

Discussed without objection on:	freebsd-www
Reviewed by:			trhodes
Approved by:			trhodes


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@20308 c2e8774f-c49f-e111-b436-862b2bbc8956