Update article to use 3 branches diagram instead of one.
Also remove branches.ascii file.  This file is highly outdated, and
I was failed to understand how it was used in article at all.  All
my attempts to get it *really* included into generated html were failed.

If someone feels that it should not be remove, feel free to re-add
it back and update it to current state of the world.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@20591 c2e8774f-c49f-e111-b436-862b2bbc8956