Fix the RTF build:
Jade is not able to embed images but links them instead, so images have
to be installed with the .rtf document (it's similar to the HTML case).

I added a new format: rtf.tar which is a better solution than using rtf
since the images are not embedded.
make FORMATS=rtf.tar is your friend :)


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@21946 c2e8774f-c49f-e111-b436-862b2bbc8956