Put on my chain mail suit and change the output format of the PR database queries once again, because I've received complaints that the HTML output mode of the PR query scripts outputs huge, bloated tables after the style changes I made to beautify HTML output.
This moves all the visual tweaks of the output in a CSS stylesheet and
add a $t_style option to the common cgi-style.pl script.  If not set,
the style defaults to empty, but users of cgi-style.pl can set this to
a local CSS style like this:

        $t_style="<style type=\"text/css\">...</style>";

or use it to refer from the header of the HTML output to an external
CSS stylesheet.

The HTML output size is substantially reduced now (at least 50% of the
size was caused by many repetitions of style="" attributes), but it's
still larger than the preformatted <pre> output because of all the
<tr> and <td> tags.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@22600 c2e8774f-c49f-e111-b436-862b2bbc8956