Remove stub chapters that are relics of Jeroen's original proposal for a Developer Handbook.
In 5 years we haven't had anyone contribute a chapter on Alpha
Assembly Language programming, Linux Compatability, Syscons, UFS,
AFS/NFS, or Signals.  If someone wants to in the future, great, but
there is no need for empty stub chapters anymore.

There are much better sources for this content in the D&I book and
Stevens' APUE.

Removing these stubs makes the content that we do have more apparent,
rather than focussing on what has yet to be written.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@24443 c2e8774f-c49f-e111-b436-862b2bbc8956