Bring in two new chapters:
Audit - not hooked up to the build, but provided for when the audit framework
becomes available in the base system.

GEOM - Incomplete, only covers striping and mirroring since that is all I
have tested.  Incomplete is better than nothing in my opinion.  This one is
hooked up to the build.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@25313 c2e8774f-c49f-e111-b436-862b2bbc8956