Add a new article, "How To Choose The Version Of FreeBSD That Is Right For You."
This article attempts to provide an introduction to such concepts
as Releases, Branches, and STABLE vs. CURRENT, which are currently
scattered around between the FAQ, various Release Engineering documents,
and folklore on the mailing lists.

The material in the FAQ should probably be deleted and this used as
its replacement.  This material does not replace any of the Release
Engineering documentation, but it does attempt to discuss how the
concepts discussed in great detail therein should influence user
decisions on which version to install.  In particular, this article
attempts to inform users about the current thinking of the development
team in terms of how future releases are going to be done.

A companion article comparing e.g. 5.X vs 6.X would be useful but is
outside the scope of this commit.

This article is partially a response to the "Quality of FreeBSD" thread
on freebsd-stable in July 2005 (and its many antecedents).

Submitted for review to:	core, re, secteam
Reviewed by:			imp, wes, remko, simon, and others


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@25389 c2e8774f-c49f-e111-b436-862b2bbc8956