Convert a string in the id attribute to the upper case to avoid confusion.
Rationale: The ISO 8879 (SGML) specification defines this as
case-insensitive and it must always be treated as one in the upper case
when the SGML declaration does not include "NAMECASE GENERAL NO."
On the other hand, in the XML specification (REC-xml-20040204)
it is case-sensitive and no normalization with regard to the case
must not be performed by an XML parser.  We use HTML 4.01 (SGML) DTD
in most of the files in the www tree and convert them into XHTML
1.0 (XML) DTD forcibly by using tidy.  So, as long as we use both
SGML and XML DTD in this way it is better to use upper case only
for the id attribute to avoid the case sensitivity problem.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@26262 c2e8774f-c49f-e111-b436-862b2bbc8956