Refactor the "confirmation code" stuff into a general purpose script.
confirm-code.cgi contains a preconfigured list of databases and their
parameters.  When a request comes in, the database in the request's 'db'
parameter is checked for validity, and a code is generated, stored in
the appropriate database and returned.

Use this new script in send-pr.sgml and remove sendpr-code.cgi which is
now superceded.


git-svn-id: svn+ssh://svn.freebsd.org/doc/www@26515 c2e8774f-c49f-e111-b436-862b2bbc8956