Fix problem marked with "XXX" in revision 1.223 by Ceri:
Add an XXX comment regarding account locking procedures.  This text is
currently wrong, but the correct way to do this is currently pending a
resolution in bin/71147.

Yar resolved this issue with his modifications to pam_unix.c revision 1.52
which was later MFCed.

The fix was to rewrite portions of this section to cover the differences between
a *'red account and a *LOCKED* account.  While here, remove the comment.

Reviewed by:	ceri


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@31328 c2e8774f-c49f-e111-b436-862b2bbc8956