Hook-up advistories and patches to be installed to   www.freebsd.org/security/{advisories,patches}/... from the en_US.ISO8859-1 directory as we do for logos, etc.
The magic is a bit special to avoid having to add each advisory
or errata notice, and their patches and signatures individually.

Factor out the slightly different to bsd.links.mk SYMLINKS
support as well as the "bulk data install" that cannot be
handled by just DATA as (i) the input dir differs, (ii) input
uses a 1 or more level of sub-directories.

This is the second of three steps.  The last one will be to link
to these files locally rather than to security.freebsd.org.

Approved by:	gabor


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@39569 c2e8774f-c49f-e111-b436-862b2bbc8956