Add two latest advisories:
Fix an integer overflow in computing the size of a temporary buffer
can result in a buffer which is too small for the requested
operation. [13:09]

Fix a bug that could lead to kernel memory disclosure with
SCTP state cookie. [13:10]

Add latest errata notices:

Fix a data corruption problem with mfi(4) operating on > 2TB
disks in a JBOD. [EN-13:03]


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@42568 c2e8774f-c49f-e111-b436-862b2bbc8956