Minimum Hardware Requirements: installation instructions
Sacrifice some technical pedanticness to simplify the description of amd64 and
i386.  For the users whom are actually confused as to which system they should
use, using more plain language (such as 32-bit vs 64-bit) and referencing the
vendors directly would be more helpful.

Also, stop mentioning that both UP and SMP are supported.  OpenBSD finished that
project in 2004.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@46400 c2e8774f-c49f-e111-b436-862b2bbc8956