kensmith@ has decided to step down from the RE role.
On behalf of the Release Engineering Team, thank you, Ken,
for your time and contributions to the team.

Ken has been a member of the Release Engineering Team since
the 4.x days, and it is of great personal sadness to see him
leave.

My hat is off to you, and all you have done for the FreeBSD
Project, and I hope you will re-join the team someday.


git-svn-id: svn+ssh://svn.freebsd.org/doc/head@46624 c2e8774f-c49f-e111-b436-862b2bbc8956