Update to r47273: Update installation chapter references, and remove references to the install chapter covering 8.x.
Update to r47233:
Be consistent about the blocksize (1m instead of 64k)
with the instructions in the release notes when describing how to
burn the memstick image to a memory stick.

Update to r46774:
Update link to the Plop boot manager.

Update to r46402:
Minimum Hardware Requirements: hardware requirements
Point people directly to the website and skip having people check for
HARDWARE.txt on media they likely haven't downloaded yet.

Update to r45127:
Remove dash from example partition label and clarify wording.  A dash
can be used in a partition label, but not in a UFS label.  Even though
users are told not to use identical labels, we can remove one source
of potential confusion by sticking to the lowest common denominator.


git-svn-id: svn+ssh://repo.freebsd.org/doc/head@49181 c2e8774f-c49f-e111-b436-862b2bbc8956