Update to r49827:
Clarify wording in the ports section of the Handbook.

The presentation assumes more knowledge on the part of a new user to
FreeBSD than it should. Some of the text, while correct, is still too
easy to misunderstand.

The overall sense of the text does not change.

Add note about ports not running their startup scripts upon installation
as being by design.


git-svn-id: svn+ssh://repo.freebsd.org/doc/head@49899 c2e8774f-c49f-e111-b436-862b2bbc8956